+++ src/main/java/com/multi/hereevent/event/EventDAOImpl.java
@@ -33,7 +33,7 @@ public class EventDAOImpl implements EventDAO{
         return sqlSession.selectList("com.multi.hereevent.event.selectAll");
     }
     @Override
-    public int deleteEvent(List<Long> eventNo) {
+    public int deleteEvent(List<Integer> eventNo) {
         return sqlSession.delete("com.multi.hereevent.event.deleteEvent", eventNo);
     }
